Split webpack into several files
