+++ src/renderer/managers/plugins.ts
@@ -2,7 +2,7 @@
 import { loadStyleSheet } from "../util";
 import type { PluginExports, RepluggedPlugin } from "../../types";
 import { Logger } from "../modules/logger";
-import { patchPlaintext } from "../modules/webpack";
+import { patchPlaintext } from "../modules/webpack/plaintext-patch";
 import { init } from "../apis/settings";
 import type { AddonSettings } from "src/types/addon";
